requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.5

# Experiment Tracking
mlflow>=2.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.21.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Data Validation
great-expectations>=0.15.0

# Logging
loguru>=0.6.0
